@@ -2,8 +2,7 @@ package com.ead.authuser.dtos;
 
 import java.util.UUID;
 
-import javax.validation.constraints.NotNull;
-
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
