@@ -4,8 +4,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-import javax.transaction.Transactional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -20,6 +18,8 @@ import com.ead.course.repositories.LessonRepositoy;
 import com.ead.course.repositories.ModuleRepository;
 import com.ead.course.repositories.UserRepository;
 import com.ead.course.services.CourseService;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class CourseServiceImpl implements CourseService {
