@@ -3,9 +3,6 @@ package com.ead.course.specifications;
 import java.util.Collection;
 import java.util.UUID;
 
-import javax.persistence.criteria.Expression;
-import javax.persistence.criteria.Root;
-
 import org.springframework.data.jpa.domain.Specification;
 
 import com.ead.course.models.CourseModel;
@@ -13,6 +10,8 @@ import com.ead.course.models.LessonModel;
 import com.ead.course.models.ModuleModel;
 import com.ead.course.models.UserModel;
 
+import jakarta.persistence.criteria.Expression;
+import jakarta.persistence.criteria.Root;
 import net.kaczmarzyk.spring.data.jpa.domain.Equal;
 import net.kaczmarzyk.spring.data.jpa.domain.Like;
 import net.kaczmarzyk.spring.data.jpa.web.annotation.And;
